setup eth rpc client trait
