+++ crml/eth-bridge/src/impls.rs
@@ -1,0 +1,136 @@
+/* Copyright 2021-2022 Centrality Investments Limited
+*
+* Licensed under the LGPL, Version 3.0 (the "License");
+* you may not use this file except in compliance with the License.
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+* You may obtain a copy of the License at the root of this project source code,
+* or at:
+*     https://centrality.ai/licenses/gplv3.txt
+*     https://centrality.ai/licenses/lgplv3.txt
+*/
+use crate::{log, types::BridgeEthereumRpcApi};
+#[cfg(not(feature = "std"))]
+use sp_std::alloc::string::ToString;
+use sp_std::{convert::TryInto, prelude::*};
+#[cfg(std)]
+use std::string::ToString;
+
+/// Provides minimal ethereum RPC queries for eth bridge protocol
+pub struct EthereumRpcClient;
+
+impl BridgeEthereumRpcApi for EthereumRpcClient {
+	/// Get transaction receipt from eth client
+	fn get_transaction_receipt(tx_hash: EthHash) -> Result<Option<TransactionReceipt>, Error<T>> {
+		let random_request_id = u32::from_be_bytes(sp_io::offchain::random_seed()[..4].try_into().unwrap());
+		let request = GetTxReceiptRequest::new(tx_hash, random_request_id as usize);
+		let resp_bytes = Self::query_eth_client(Some(request)).map_err(|e| {
+			log!(error, "💎 read eth-rpc API error: {:?}", e);
+			<Error<T>>::HttpFetch
+		})?;
+
+		let resp_str = core::str::from_utf8(&resp_bytes).map_err(|_| {
+			log!(error, "💎 response invalid utf8: {:?}", resp_bytes);
+			<Error<T>>::HttpFetch
+		})?;
+
+		// Deserialize JSON to struct
+		serde_json::from_str::<EthResponse<TransactionReceipt>>(resp_str)
+			.map(|resp| resp.result)
+			.map_err(|err| {
+				log!(error, "💎 deserialize json response error: {:?}", err);
+				<Error<T>>::HttpFetch
+			})
+	}
+
+	/// Get latest block number from eth client
+	fn get_block(req: LatestOrNumber) -> Result<Option<EthBlock>, Error<T>> {
+		let request = match req {
+			LatestOrNumber::Latest => GetBlockRequest::latest(1_usize),
+			LatestOrNumber::Number(n) => GetBlockRequest::for_number(1_usize, n),
+		};
+		let resp_bytes = Self::query_eth_client(request).map_err(|e| {
+			log!(error, "💎 read eth-rpc API error: {:?}", e);
+			<Error<T>>::HttpFetch
+		})?;
+
+		let resp_str = core::str::from_utf8(&resp_bytes).map_err(|_| {
+			log!(error, "💎 response invalid utf8: {:?}", resp_bytes);
+			<Error<T>>::HttpFetch
+		})?;
+
+		// Deserialize JSON to struct
+		serde_json::from_str::<EthResponse<EthBlock>>(resp_str)
+			.map(|resp| resp.result)
+			.map_err(|err| {
+				log!(error, "💎 deserialize json response error: {:?}", err);
+				<Error<T>>::HttpFetch
+			})
+	}
+
+	/// This function uses the `offchain::http` API to query the remote ethereum information,
+	/// and returns the JSON response as vector of bytes.
+	fn query_eth_client<R: serde::Serialize>(request_body: R) -> Result<Vec<u8>, Error<T>> {
+		// Load eth http URI from offchain storage
+		// this should have been configured on start up by passing e.g. `--eth-http`
+		// e.g. `--eth-http=http://localhost:8545`
+		let eth_http_uri = if let Some(value) = sp_io::offchain::local_storage_get(StorageKind::PERSISTENT, b"ETH_HTTP")
+		{
+			value
+		} else {
+			log!(
+				error,
+				"💎 Eth http uri is not configured! set --eth-http=<value> on start up"
+			);
+			return Err(Error::<T>::OcwConfig);
+		};
+		let eth_http_uri = core::str::from_utf8(&eth_http_uri).map_err(|_| Error::<T>::OcwConfig)?;
+
+		const HEADER_CONTENT_TYPE: &str = "application/json";
+		log!(info, "💎 sending request to: {}", eth_http_uri);
+		let body = serde_json::to_string::<R>(&request_body).unwrap();
+		let body_raw = body.as_bytes();
+		// Initiate an external HTTP GET request. This is using high-level wrappers from `sp_runtime`.
+		let request = rt_offchain::http::Request::post(eth_http_uri, vec![body_raw]);
+		log!(trace, "💎 request: {:?}", request);
+
+		// Keeping the offchain worker execution time reasonable, so limiting the call to be within 3s.
+		let timeout = sp_io::offchain::timestamp().add(rt_offchain::Duration::from_millis(REQUEST_TTL_MS));
+		let pending = request
+			.add_header("Content-Type", HEADER_CONTENT_TYPE)
+			.add_header("Content-Length", &body_raw.len().to_string())
+			.deadline(timeout) // Setting the timeout time
+			.send() // Sending the request out by the host
+			.map_err(|err| {
+				log!(error, "💎 http request error: {:?}", err);
+				<Error<T>>::HttpFetch
+			})?;
+
+		// By default, the http request is async from the runtime perspective. So we are asking the
+		// runtime to wait here.
+		// The returning value here is a `Result` of `Result`, so we are unwrapping it twice by two `?`
+		// ref: https://substrate.dev/rustdocs/v3.0.0/sp_runtime/offchain/http/struct.PendingRequest.html#method.try_wait
+		let response = pending
+			.try_wait(timeout)
+			.map_err(|err| {
+				log!(error, "💎 http request error: timeline reached: {:?}", err);
+				<Error<T>>::HttpFetch
+			})?
+			.map_err(|err| {
+				log!(error, "💎 http request error: timeline reached: {:?}", err);
+				<Error<T>>::HttpFetch
+			})?;
+		log!(trace, "💎 response: {:?}", response);
+
+		if response.code != 200 {
+			log!(error, "💎 http request status code: {}", response.code);
+			return Err(<Error<T>>::HttpFetch);
+		}
+
+		// Read the response body and check it's valid utf-8
+		Ok(response.body().collect::<Vec<u8>>())
+	}
+}
